Use non-encoded workspace URI string in getRelativeModelUri
The getRelativeModelUri receives as a parameter a non-encoded string.
Use the non-encoded workspace URI string when resolving relative URIs.
This is needed as otherwise the string is url encoded and thus
characters like ' ' and ':' are encoded. This is especially important
when handling Windows paths as those contain ':'.

Contributed on behalf of STMicroelectronics